+++ .gitignore
@@ -1,7 +1,7 @@
 .DS_Store
 .eslintcache
 .idea
-/.vscode
+.vscode
 coverage
 dist
 node_modules
