+++ lib/getBaseConfig.js
@@ -1,11 +1,11 @@
-const { commands } = require('./constants');
+const presets = require('./presets');
 
 module.exports = (commandName) => {
-  if (commands[commandName] === undefined) {
+  if (presets[commandName] === undefined) {
     throw new Error(`fedx-scripts: ${commandName} is unsupported in this version`);
   }
 
-  const { configFile } = commands[commandName];
+  const { configFile } = presets[commandName];
   const config = require(configFile);
   return config;
 };
