+++ lib/getProjectConfigFile.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const glob = require('glob');
 
-const { commands } = require('./constants');
+const presets = require('./presets');
 const { APP_ROOT } = require('./paths');
 
 const getNoProjectConfigMessage = (commandName, suggestedName) => `
@@ -40,7 +40,7 @@ module.exports = (commandName) => {
     globSearchPattern,
     suggestedName,
     configFile,
-  } = commands[commandName];
+  } = presets[commandName];
   const globOptions = {
     cwd: APP_ROOT,
     nodir: true,
